week2/day6/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.9
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.1
